requirements.txt
certifi==2025.4.26
charset-normalizer==3.4.2
greenlet==3.2.1
idna==3.10
numpy==2.2.5
pandas==2.2.3
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
requests==2.32.3
six==1.17.0
SQLAlchemy==1.4.46
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
apache-airflow==2.9.0
apache-airflow-providers-postgres==2.4.0